Rename v1-icon-builder spec to proper test naming convention
